# infrastructure/terraform/gke-cluster/openapi-copy-2.yaml
swagger: "2.0"
info:
  title: MWC Microservice API
  version: "1.0"
host: "api.fadellh.com"
schemes:
  - https
paths:
...
  "/v1/users/verify":
    options:
      summary: "CORS preflight request"
      operationId: optionsVerifyCustomer
      responses:
        "200":
          description: "CORS headers"
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
      x-google-backend:
        address: "https://backend.fadellh.com"
        path_translation: APPEND_PATH_TO_ADDRESS

    post:
      summary: "Verify a user by email"
      operationId: verifyCustomer
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: "VerifyCustomerCommand data"
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
      responses:
        "200":
          description: "User verified successfully"
        "404":
          description: "User not found"
      security:
        - api_key: []
      x-google-backend:
        address: "https://backend.fadellh.com"
        path_translation: APPEND_PATH_TO_ADDRESS
        jwt_audience: "dti-purwa"
  "/v1/users/create":
    options:
      summary: "CORS preflight request"
      operationId: optionsCreateCustomer
      responses:
        "200":
          description: "CORS headers"
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
      x-google-backend:
        address: "https://backend.fadellh.com"
        path_translation: APPEND_PATH_TO_ADDRESS

    post:
      summary: "Create a new user"
      operationId: createCustomer
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: "CreateCustomerCommand data"
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              fullname:
                type: string
      responses:
        "200":
          description: "User created (or returned if already exists)"
        "400":
          description: "Bad Request"
      security:
        - api_key: []
      x-google-backend:
        address: "https://backend.fadellh.com"
        path_translation: APPEND_PATH_TO_ADDRESS
        jwt_audience: "dti-purwa"

############################################################
#  SECURITY DEFINITIONS
############################################################
securityDefinitions:
  api_key:
    type: apiKey
    name: key
    in: query
  firebase_jwt:
    type: "oauth2"
    flow: "implicit"
    authorizationUrl: ""
    x-google-issuer: "https://securetoken.google.com/dti-purwa"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/jwk/securetoken@system.gserviceaccount.com"
    x-google-audiences: "dti-purwa"